+++ examples/external-data/jest.config.cjs
@@ -34,4 +34,5 @@ module.exports = {
 			},
 		],
 	],
+	testTimeout: 10000,
 };