+++ examples/external-data/src/mock-customer-service/service.ts
@@ -7,7 +7,6 @@ import { Server } from "node:http";
 
 import cors from "cors";
 import express from "express";
-import fetch from "node-fetch";
 
 import { ITaskData, assertValidTaskData } from "../model-interface/index.js";
 import { ClientManager } from "../utilities/index.js";