+++ packages/dds/merge-tree/src/test/sortedSegmentSet.spec.ts
@@ -5,6 +5,8 @@
 
 import { strict as assert } from "node:assert";
 
+import { makeRandom } from "@fluid-private/stochastic-test-utils";
+
 import { LocalReferencePosition } from "../localReference.js";
 import { ISegment, type ISegmentPrivate } from "../mergeTreeNodes.js";
 import { TrackingGroup } from "../mergeTreeTracking.js";
@@ -90,6 +92,41 @@ describe("SortedSegmentSet", () => {
 		}
 		assert.equal(set.size, segmentCount);
 		validateSet(client, set, (i) => toMergeNodeInfo(i)?.ordinal);
+	});
+
+	describe("SortedSegmentSet of local references", () => {
+		it("Inserts in order", () => {
+			const random = makeRandom(0);
+			const refsAtAllPositions: LocalReferencePosition[] = [];
+			for (let i = 0; i < client.getLength(); i++) {
+				const { segment, offset } = client.getContainingSegment<ISegmentPrivate>(i);
+				assert(segment !== undefined);
+				assert(offset !== undefined);
+				refsAtAllPositions.push(
+					client.createLocalReferencePosition(
+						segment,
+						offset,
+						ReferenceType.SlideOnRemove,
+						undefined,
+					),
+				);
+			}
+
+			random.shuffle(refsAtAllPositions);
+			const segmentSet = new SortedSegmentSet<LocalReferencePosition>();
+			for (const ref of refsAtAllPositions) {
+				segmentSet.addOrUpdate(ref);
+			}
+
+			// Validate that the set is sorted
+			const refsBackToPositions = segmentSet.items.map((ref) =>
+				client.localReferencePositionToPosition(ref),
+			);
+			assert.deepEqual(
+				refsBackToPositions,
+				Array.from({ length: client.getLength() }, (_, i) => i),
+			);
+		});
 	});
 
 	it("SortedSegmentSet of local references", () => {