+++ tools/api-markdown-documenter/src/api-item-transforms/default-implementations/TransformApiTypeLike.ts
@@ -5,7 +5,6 @@
 
 import {
 	type ApiCallSignature,
-	type ApiClass,
 	type ApiConstructor,
 	type ApiIndexSignature,
 	type ApiItem,
@@ -20,13 +19,14 @@ import {
 	getApiItemKind,
 	getScopedMemberNameForDiagnostics,
 	isStatic,
+	type ApiTypeLike,
 } from "../../utilities/index.js";
 import { getFilteredMembers } from "../ApiItemTransformUtilities.js";
 import type { ApiItemTransformationConfiguration } from "../configuration/index.js";
 import { createChildDetailsSection, createMemberTables } from "../helpers/index.js";
 
 /**
- * Default documentation transform for `Class` items.
+ * Default documentation transform for {@link ApiTypeLike | type-like} API items.
  *
  * @remarks Format:
  *
@@ -64,14 +64,14 @@ import { createChildDetailsSection, createMemberTables } from "../helpers/index.
  *
  * - index-signatures
  */
-export function transformApiClass(
-	apiClass: ApiClass,
+export function transformApiTypeLike(
+	apiItem: ApiTypeLike,
 	config: ApiItemTransformationConfiguration,
 	generateChildContent: (apiItem: ApiItem) => SectionNode[],
 ): SectionNode[] {
 	const sections: SectionNode[] = [];
 
-	const filteredChildren = getFilteredMembers(apiClass, config);
+	const filteredChildren = getFilteredMembers(apiItem, config);
 	if (filteredChildren.length > 0) {
 		// Accumulate child items
 		const constructors: ApiConstructor[] = [];
@@ -82,11 +82,13 @@ export function transformApiClass(
 		for (const child of filteredChildren) {
 			const childKind = getApiItemKind(child);
 			switch (childKind) {
-				case ApiItemKind.Constructor: {
+				case ApiItemKind.Constructor:
+				case ApiItemKind.ConstructSignature: {
 					constructors.push(child as ApiConstructor);
 					break;
 				}
-				case ApiItemKind.Property: {
+				case ApiItemKind.Property:
+				case ApiItemKind.PropertySignature: {
 					allProperties.push(child as ApiPropertyItem);
 					break;
 				}
@@ -98,16 +100,17 @@ export function transformApiClass(
 					indexSignatures.push(child as ApiIndexSignature);
 					break;
 				}
-				case ApiItemKind.Method: {
+				case ApiItemKind.Method:
+				case ApiItemKind.MethodSignature: {
 					allMethods.push(child as ApiMethod);
 					break;
 				}
 				default: {
 					config.logger?.error(
-						`Child item "${
-							child.displayName
-						}" of Class "${getScopedMemberNameForDiagnostics(
-							apiClass,
+						`Child item "${child.displayName}" of ${
+							apiItem.kind
+						} "${getScopedMemberNameForDiagnostics(
+							apiItem,
 						)}" is of unsupported API item kind: "${childKind}"`,
 					);
 					break;
@@ -256,5 +259,5 @@ export function transformApiClass(
 		}
 	}
 
-	return config.defaultSectionLayout(apiClass, sections, config);
+	return config.defaultSectionLayout(apiItem, sections, config);
 }